Added consistency test and improved sumsq definition
Added a consistency test to avoid unrealistic results where Ratm or Rw*T
exceed Rprime_noglint, the TOA reflectance corrected for rayleigh but
not glint.

Also improved the definition of the cost function by normalizing by the
sum of weights.
This has an impact on backscattering constraint for clear waters and
should improve the inter-sensor consistency over clear waters.
The amplitude of the constraint is modified accordingly.

Updated parameters to reproduce v3.5 results;